+++ system/lib/wasmfs/file.h
@@ -101,9 +101,8 @@ protected:
   // A mutex is needed for multiple accesses to the same file.
   std::recursive_mutex mutex;
 
-  // The the size in bytes of a file or return a negative error code. May be
-  // called on files that have not been opened.
-  virtual off_t getSize() = 0;
+  // May be called on files that have not been opened.
+  virtual size_t getSize() = 0;
 
   mode_t mode = 0; // User and group mode bits for access permission.
 
@@ -145,7 +144,7 @@ protected:
   // Sets the size of the file to a specific size. If new space is allocated, it
   // should be zero-initialized. May be called on files that have not been
   // opened. Returns 0 on success or a negative error code.
-  virtual int setSize(off_t size) = 0;
+  virtual int setSize(size_t size) = 0;
 
   // Sync the file data to the underlying persistent storage, if any. Returns 0
   // on success or a negative error code.
@@ -255,7 +254,7 @@ public:
 protected:
   // 4096 bytes is the size of a block in ext4.
   // This value was also copied from the JS file system.
-  off_t getSize() override { return 4096; }
+  size_t getSize() override { return 4096; }
 };
 
 class Symlink : public File {
@@ -271,7 +270,7 @@ public:
   virtual std::string getTarget() const = 0;
 
 protected:
-  off_t getSize() override { return getTarget().size(); }
+  size_t getSize() override { return getTarget().size(); }
 };
 
 class File::Handle {
@@ -287,7 +286,7 @@ public:
   Handle(std::shared_ptr<File> file) : file(file), lock(file->mutex) {}
   Handle(std::shared_ptr<File> file, std::defer_lock_t)
     : file(file), lock(file->mutex, std::defer_lock) {}
-  off_t getSize() { return file->getSize(); }
+  size_t getSize() { return file->getSize(); }
   mode_t getMode() { return file->mode; }
   void setMode(mode_t mode) {
     // The type bits can never be changed (whether something is a file or a
@@ -326,7 +325,7 @@ public:
     return getFile()->write(buf, len, offset);
   }
 
-  [[nodiscard]] int setSize(off_t size) { return getFile()->setSize(size); }
+  [[nodiscard]] int setSize(size_t size) { return getFile()->setSize(size); }
 
   // TODO: Design a proper API for flushing files.
   [[nodiscard]] int flush() { return getFile()->flush(); }
